#7.4 Protecting Video Routes

+++ controllers/videoController.js
@@ -100,10 +100,14 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", {
-            pageTitle: `Edit ${video.title}`,
-            video,
-        });
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            res.render("editVideo", {
+                pageTitle: `Edit ${video.title}`,
+                video,
+            });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
@@ -139,9 +143,14 @@ export const deleteVideo = async (req, res) => {
         },
     } = req;
     try {
-        await Video.findOneAndRemove({
-            _id: id,
-        });
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        } else {
+            await Video.findOneAndRemove({
+                _id: id,
+            });
+        }
     } catch (error) {
         console.log(error);
     }
